Updated all tests to work with recent changes

+++ test/unit_testing/tests/Map/intended/include/haxe_ds_IntMap.h
@@ -41,11 +41,11 @@ public:
 	}
 	
 	bool exists(int key) {
-		return this->m.count(key) > 0;
+		return (int)(this->m.count(key)) > 0;
 	}
 	
 	bool remove(int key) {
-		return this->m.erase(key) > 0;
+		return (int)(this->m.erase(key)) > 0;
 	}
 	
 	std::shared_ptr<Iterator<int>> keys() {
